requirements.txt
azure-functions==1.20.0
pytest==8.2.2
ruff==0.5.7
